+++ CHANGES.rst
@@ -9,6 +9,7 @@ The Products.MeetingCommunes version must be the same as the Products.PloneMeeti
 - When applying 'meetingadvicefinances_add_advicecreated_state' WF adaptation, set advicecreated state as new_initial_state
 - In query_itemstocontrolcompletenessof, do not use the config.FINANCE_WAITING_ADVICES_STATES but compute the states in which advice
   can be given by finances groups
+- Override MeetingItem._adviceTypesForAdviser to manage finances specific advice types
 
 4.1rc6 (2019-07-02)
 -------------------